+++ src/main/java/encore/server/domain/review/converter/ReviewConverter.java
@@ -1,15 +1,11 @@
 package encore.server.domain.review.converter;
 
 import encore.server.domain.review.dto.request.ReviewDataReq;
-import encore.server.domain.review.dto.response.*;
 import encore.server.domain.review.entity.ReviewData;
 import encore.server.domain.review.entity.ReviewTags;
-import encore.server.domain.review.entity.ViewImage;
-import encore.server.domain.review.enumerate.LikeType;
 import encore.server.domain.review.enumerate.Tag;
 import encore.server.domain.review.dto.request.ReviewReq;
 import encore.server.domain.review.entity.Review;
-import encore.server.domain.ticket.entity.Actor;
 import encore.server.domain.ticket.entity.Ticket;
 import encore.server.domain.user.entity.User;
 import encore.server.global.exception.ApplicationException;
@@ -32,35 +28,6 @@ public class ReviewConverter {
         return review;
     }
 
-    public static ViewImageRes toViewImageRes(List<ViewImage> viewImages) {
-        return ViewImageRes.builder()
-                .viewImages(viewImages.stream()
-                        .map(viewImage -> ViewImageRes.ViewImage.builder()
-                                .id(viewImage.getId())
-                                .url(viewImage.getUrl())
-                                .level(viewImage.getLevel())
-                                .build())
-                        .collect(Collectors.toList()))
-                .build();
-    }
-
-    public static ReviewDetailRes toReviewDetailRes(Review review, Boolean isUnlocked, LikeType likeType, String elapsedTime) {
-        return ReviewDetailRes.builder()
-                .reviewId(review.getId())
-                .ticket(toTicketRes(review.getTicket()))
-                .userId(review.getUser().getId())
-                .profileImageUrl(review.getUser().getProfileImageUrl())
-                .title(review.getTitle())
-                .tags(tagToString(review.getTags()))
-                .reviewDataRes(toReviewDataRes(review.getReviewData()))
-                .isUnlocked(isUnlocked)
-                .isMyReview(review.getUser().getId().equals(review.getUser().getId()))
-                .viewCount(review.getViewCount())
-                .elapsedTime(elapsedTime)
-                .likeRes(ReviewLikeRes.of(likeType, review))
-                .build();
-    }
-
     private static List<ReviewTags> toReviewTags(List<String> tags, Review review) {
         try {
             return tags.stream()
@@ -72,27 +39,6 @@ public class ReviewConverter {
         } catch (IllegalArgumentException e) {
             throw new ApplicationException(ErrorCode.REVIEW_TAG_NOT_FOUND_EXCEPTION);
         }
-    }
-
-    private static List<String> tagToString(List<ReviewTags> tags) {
-        return tags.stream()
-                .map(reviewTag -> reviewTag.getTag().name())
-                .collect(Collectors.toList());
-    }
-
-    //Todo: 티켓 완성 시 패키지 옮기기
-    private static ReviewDetailRes.TicketRes toTicketRes(Ticket ticket) {
-        return ReviewDetailRes.TicketRes.builder()
-                .ticketId(ticket.getId())
-                .ticketTitle(ticket.getTitle())
-                .seat(ticket.getSeat())
-                .viewedDate(ticket.getViewedDate())
-                .imageUrl(ticket.getTicketImageUrl())
-                .actors(ticket.getActors()
-                        .stream()
-                        .map(Actor::getName)
-                        .collect(Collectors.toList()))
-                .build();
     }
 
     private static ReviewData toReviewData(ReviewDataReq req){
@@ -117,32 +63,6 @@ public class ReviewConverter {
                         .performanceRating(req.rating().performanceRating())
                         .totalRating(req.rating().totalRating())
                         .ratingReview(req.rating().ratingReview())
-                        .build())
-                .build();
-    }
-
-    public static ReviewDataRes toReviewDataRes(ReviewData reviewData) {
-        return ReviewDataRes.builder()
-                .view(ReviewDataRes.View.builder()
-                        .viewLevel(reviewData.getView().getViewLevel())
-                        .viewReview(reviewData.getView().getViewReview())
-                        .build())
-                .sound(ReviewDataRes.Sound.builder()
-                        .soundLevel(reviewData.getSound().getSoundLevel())
-                        .soundReview(reviewData.getSound().getSoundReview())
-                        .build())
-                .facility(ReviewDataRes.Facility.builder()
-                        .facilityLevel(reviewData.getFacility().getFacilityLevel())
-                        .facilityReview(reviewData.getFacility().getFacilityReview())
-                        .build())
-                .rating(ReviewDataRes.Rating.builder()
-                        .numberRating(reviewData.getRating().getNumberRating())
-                        .storyRating(reviewData.getRating().getStoryRating())
-                        .revisitRating(reviewData.getRating().getRevisitRating())
-                        .actorRating(reviewData.getRating().getActorRating())
-                        .performanceRating(reviewData.getRating().getPerformanceRating())
-                        .totalRating(reviewData.getRating().getTotalRating())
-                        .ratingReview(reviewData.getRating().getRatingReview())
                         .build())
                 .build();
     }
