+++ src/main/java/encore/server/domain/review/dto/response/ViewImageRes.java
@@ -6,6 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Builder
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
@@ -23,5 +24,17 @@ public record ViewImageRes(
             @Schema(description = "시야 레벨", example = "3")
             Long level
     ) {
+    }
+
+    public static ViewImageRes of(List<encore.server.domain.review.entity.ViewImage> viewImages) {
+        return ViewImageRes.builder()
+                .viewImages(viewImages.stream()
+                        .map(viewImage -> ViewImageRes.ViewImage.builder()
+                                .id(viewImage.getId())
+                                .url(viewImage.getUrl())
+                                .level(viewImage.getLevel())
+                                .build())
+                        .collect(Collectors.toList()))
+                .build();
     }
 }
